# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Healthcare Staffing Platform API Doc
  description: |
    On-Demand Healthcare Staffing & Visit Management Portal — REST API.

    ## Authentication
    All protected endpoints require a **Bearer JWT** in the `Authorization` header:
    ```
    Authorization: Bearer <accessToken>
    ```
    Obtain tokens via `POST /auth/login`. Access tokens expire in **15 minutes**;
    use `POST /auth/refresh` with a valid refresh token to rotate.

    ## Roles
    | Role | Description |
    |------|-------------|
    | `SUPER_ADMIN` | Full platform control |
    | `RECRUITER` | Read-only admin access |
    | `FACILITY_ADMIN` | Manages one facility tenant |
    | `TEAM_MEMBER` | Scoped facility operator |
    | `NURSE` | Field professional (mobile) |

    ## File Uploads
    All uploads are stored in **DigitalOcean Spaces** (private).
    Use `POST /storage/signed-upload-url` for direct client-to-Spaces uploads,
    or `POST /storage/upload/{folder}` for server-proxied multipart uploads.
    Download links are short-lived signed URLs — never public.

    ## Pagination
    List endpoints accept `page` and `limit` query params and return:
    ```json
    { "pagination": { "page": 1, "limit": 20, "total": 100, "totalPages": 5 } }
    ```

    ## EVV Geofencing
    Check-in/out enforces a **200-metre geofence** around the case address (Haversine).
    Breaches create a `FLAGGED` visit requiring admin override approval.
  version: 1.0.0
  contact:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Reusable Parameters ──────────────────
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
...
  responses:
...
    Forbidden:
      description: Authenticated but insufficient role/scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationErrorResponse' }
...
  schemas:
...
    VisitType:
      type: string
      enum: [ADMISSION, REGULAR, RESUMPTION_OF_CARE, RECERTIFICATION,
             SUPERVISORY, DISCHARGE]
...
    OasisType:
      type: string
      enum: [ADMISSION, RESUMPTION_OF_CARE, RECERTIFICATION, DISCHARGE]
...
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
        data:    { type: object }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: 'Resource not found' }
    ValidationErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: 'Validation failed' }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    PaginationMeta:
      type: object
      properties:
        page:       { type: integer }
        limit:      { type: integer }
        total:      { type: integer }
        totalPages: { type: integer }
        hasNext:    { type: boolean }
        hasPrev:    { type: boolean }

    # ── Auth ───────────────────────────────
...
    CaseSummary:
      type: object
      properties:
        id:               { type: string, format: uuid }
        publicIdentifier: { type: string, example: 'Case-PT-7701' }
        facilityId:       { type: string, format: uuid }
        visitType:        { $ref: '#/components/schemas/VisitType' }
        city:             { type: string }
        state:            { type: string }
        isOasisCase:      { type: boolean }
        isActive:         { type: boolean }
        specialties:
          type: array
          items: { type: string }
        createdAt:        { type: string, format: date-time }
    CaseDetail:
      allOf:
        - $ref: '#/components/schemas/CaseSummary'
        - type: object
          properties:
            patientFirstName: { type: string, nullable: true }
            patientLastName:  { type: string, nullable: true }
            dateOfBirth:      { type: string, format: date, nullable: true }
            gender:           { type: string, nullable: true }
            primaryDiagnosis: { type: string, nullable: true }
            addressLine1:     { type: string }
            addressLine2:     { type: string, nullable: true }
            zipCode:          { type: string }
            latitude:         { type: number, nullable: true }
            longitude:        { type: number, nullable: true }
            isOasisCase:      { type: boolean }
            oasisType:        { $ref: '#/components/schemas/OasisType', nullable: true }
            notes:            { type: string, nullable: true }
            facility:
              type: object
              properties:
                id:   { type: string, format: uuid }
                name: { type: string }

    # ── Shifts ─────────────────────────────
...
security:
  - BearerAuth: []

# docs/paths-part1.yaml  (auth · users · credentials · facilities)
paths:
...
  /cases:
    post:
      tags: [Cases]
      summary: Create a patient case
      description: |
        **Roles:** SUPER_ADMIN, FACILITY_ADMIN, TEAM_MEMBER  
        A random `publicIdentifier` (e.g. `Case-PT-7701`) is auto-assigned to protect PHI.
        Nurses never see real patient names.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [facilityId, addressLine1, city, state, zipCode, visitType]
              properties:
                facilityId:       { type: string, format: uuid }
                patientFirstName: { type: string }
                patientLastName:  { type: string }
                dateOfBirth:      { type: string, format: date }
                gender:           { type: string }
                primaryDiagnosis: { type: string }
                notes:            { type: string }
                addressLine1:     { type: string }
                addressLine2:     { type: string }
                city:             { type: string }
                state:            { type: string }
                zipCode:          { type: string }
                visitType:        { $ref: '#/components/schemas/VisitType' }
                specialties:
                  type: array
                  items: { type: string }
                  example: [WOUND_CARE, PICC_LINE]
                isOasisCase:      { type: boolean }
                oasisType:        { $ref: '#/components/schemas/OasisType' }
      responses:
        '201':
          description: Case created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/CaseDetail' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }

    get:
      tags: [Cases]
      summary: List cases
      description: |
        Facility users see only their facility's cases.
        Nurses see a filtered list (patient PII masked).
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: facilityId
          in: query
          schema: { type: string, format: uuid }
        - name: visitType
          in: query
          schema: { $ref: '#/components/schemas/VisitType' }
        - name: isActive
          in: query
          schema: { type: boolean }
        - name: search
          in: query
          description: Search by public identifier or patient name
          schema: { type: string }
      responses:
        '200':
          description: Paginated case list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:    { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/CaseSummary' }
                  pagination: { $ref: '#/components/schemas/PaginationMeta' }
  /cases/{id}:
    get:
      tags: [Cases]
      summary: Get case detail
      description: Patient PII is redacted for NURSE role.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Case detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/CaseDetail' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      tags: [Cases]
      summary: Update a case
      description: '**Roles:** SUPER_ADMIN, FACILITY_ADMIN, TEAM_MEMBER (own facility)'
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                patientFirstName: { type: string }
                patientLastName:  { type: string }
                primaryDiagnosis: { type: string }
                notes:            { type: string }
                addressLine1:     { type: string }
                city:             { type: string }
                state:            { type: string }
                zipCode:          { type: string }
                visitType:        { $ref: '#/components/schemas/VisitType' }
                specialties:
                  type: array
                  items: { type: string }
                isOasisCase:      { type: boolean }
                oasisType:        { $ref: '#/components/schemas/OasisType' }
                isActive:         { type: boolean }
      responses:
        '200':
          description: Case updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/CaseDetail' }

    delete:
      tags: [Cases]
      summary: Archive (soft-delete) a case
      description: '**Roles:** SUPER_ADMIN, FACILITY_ADMIN'
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Case archived
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }

  # ═══════════════════════════════════════════
  # SHIFTS
  # ═══════════════════════════════════════════
...